ThermalDetector/requirements.txt
bleak==0.14.3
joblib==1.2.0
keras==2.10.0
matplotlib==3.6.0
numpy==1.22.3
paho_mqtt==1.6.1
pandas==1.4.2
PyYAML==6.0
ruamel.base==1.0.0
scikit_learn==1.1.2
seaborn==0.11.2
tensorflow==2.10.0